Created index.ts file for routes
