+++ Week 3/Job Portal Management/package.json
@@ -4,9 +4,9 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "dev": "nodemon --exec \"ts-node --esm\" src/index.ts",
+    "dev": "nodemon --exec \"ts-node --esm\" src/server.ts",
     "build": "tsc",
-    "start": "node dist/index.js"
+    "start": "node dist/server.js"
   },
   "keywords": [],
   "author": "",
